Now we use double for square size, so when it comes to a large grids we still draw it accurately

+++ src/input/MyMouseListener.java
@@ -42,7 +42,7 @@ public class MyMouseListener extends MouseAdapter {
             renderer.setLastMouseX(e.getX());
             renderer.setLastMouseY(e.getY());
         } else {
-            int squareSize = renderer.calcSquareSize();
+            double squareSize = renderer.calcSquareSize();
 
             int x = (int) ((e.getX() + renderer.getPanOffsetX()) / squareSize);
             int y = (int) ((e.getY() + renderer.getPanOffsetY() - 40) / squareSize);
